Add volume controls
Adds two sliders for volume. The primary one is below the
player. Another one is in the new settings panel that can be accessed
by clicking a button. This one sets the "max volume".

The actual volume is calculated by multiplying the two. This allows
the user to have better control when setting a lower volume.

Both volume values are stored in localStorage so they persist over
sessions.